requirements.txt
pyserial
evdev
